requirements.txt
rumps
pyobjc-framework-Cocoa
pyobjc-framework-Quartz
sounddevice
numpy
websockets
python-dotenv
